gooddata-pandas/test-requirements.txt
pytest~=6.2.4
pytest-cov~=2.12.1
vcrpy~=4.1.1
python-dotenv~=0.19.0
